[Feat] #38 - 데이터 불러오기 완료 시 뷰 띄우기
