+++ Diptych/Diptych/Presentation/TodayDiptych/Model/WeeklyData.swift
@@ -1,0 +1,13 @@
+//
+//  WeeklyData.swift
+//  Diptych
+//
+//  Created by 김민 on 2023/07/23.
+//
+
+import Foundation
+
+struct WeeklyData {
+    let diptychState: DiptychState
+    let thumbnail: String?
+}
